BGDIINF_SB-1702: Added text-danger to details icon and added correct animation
Now when the layer details is shown the cog icon is red due to class
text-danger.

Also added the pulse animation on the hidden transition, before the
pulse animation was only on the show transition. For this to simplify
the code we use a new external library https://jofftiquez.github.io/v-animate-css/.