BGDIINF_SB-2089: Fixed drawing measurement
The drawing measurement click and double click action were not
triggered, which means that we could not end the measurement drawing
tool nor add subsequent measurement point.

This was due to the fact that the measurement tooltip was over the
drawing measurement and that this overlay did not pass the click event
futher. Moving the tooltip away from the measurement line would also
have fixed the issue but the display wouldn't have been niche (also not
sure if this would have worked on a touch display).

Also center the measurement tooltip on the line, it was one pixel off.