BGDIINF_SB-1993: Fixed issue with legacy KML adminid param
The legacy KML adminid param did not work because of an exception not
handled correctly. In case of this exception the `routeChangeIsTriggeredByThisModule`
flag was not reseted which means that subsequent event trigered by the
asynchrone legacy adminid resolution would not be treated due to this
flag.

So now in case of a failure to re-route, clear the local flag to make
sure to treat any subsequent events.