BGDIINF_SB-2021: Clear the drawing on removing the layer
When removing a drawing layer, the drawing is cleared allow the creation
of a new drawing. Note the actual drawing file on the backend is not
touched ! This only clear the drawing from the frontend module and the
drawing can still be retrieved via its ID from the backend.

For this I needed to add a new event in the drawing manager to
differenciate the clear event from other change event. The clear event
would then only trigger a kml update when the drawing mode is active
(drawing menu open). This avoid creating empty kml by removing a drawing
layer.