BGDIINF_SB-2114: Fixed drawing delete button
The delete button in the drawing menu was always deactivated.

This was due to the fact that vue computed properties are now (with Vue
3.x at least was maybe different on Vue 2.x) only evaluated if they
depends on reactive data. Because drawingNotEmpty did not depends on
reactive data but only on the external DrawingManager with is pure js,
this flag was never re-evaluated.

Now we use a method that is called each time a drawing is changed.