BGDIINF_SB-2114: Fixed footer with portal
Vue 3.x has a new teleport feature that is meant to replace portal-vue.
However this has some limitations, for instance the teleport destination
has to be mounted before the teleport and for multi teleport the
sequence is defined by the teleport sequence. Without reworking the
Footer we cannot use Vue 3x teleport (see BGDIINF_SB-2119).

Therefore use the last beta version of portal-vue which supports Vue 3.