BGDIINF_SB-2114: Vue3 preparation: fixed brocken drawing module
This is a preparation for Vue 3 migration.

With Vue 3.x the whole drawing module did not worked anymore. Strange effects when
trying to draw.

The problem was due to the new Vue 3 reactivity. All object that are
parts of the Vue data property are reactive, which means that they are
proxied. Vue 3 use a deep reactive system, which means that child object
are also made reactive. One issue in the drawing module was that the
drawing manager created a new layer (not reactive) and set it to the
openlayer map (which was reactive). Latter the openlayer logic was
trying to match a layer config from drawing manager (non reactive) to a
the actal layer on the map. Because the map was reactive all layer (as
well as all its objects) were automatically reactive.

To avoid this we don't set the openlayer map object to the Vue data
property. In general it is a bad idea to set a Vue data property with a
complex object from an external library. In our case it created a nasty
bug, but in all case it has a performance penalty, therefore it is
better to avoid it. That's why I changed everywhere in the code where I
found that complex object where set in the Vue data property and set
them in the created or beforeCreate Vue hook.

Also this is not an issue with Vue 2.x it is not a good practice.