BGDIINF_SB-2114: Fixed navigation
Currently the navigation links (e.g. the one from the share drawing) are
made with `/#` so we need to use Web Hash History. We might in future
change it back to Web history but would then need to adapt all links
generated by the application, e.g. qrcode and drawing.