BGDIINF_SB-2114: Fixed setting layer config, url opacity, add/remove layer
Setting layer config (upon app start and during changing language) did
not re-add the actual active layers due to calling the wrong vuex
method.

Changing the layer opacity through url did not work due to some typo in
the logic.

Also adding/removing layer were not reflected in the url because we used
the vuex action name instead of mutation name in the LayerParamConfig as
mutation to watch.