BGDIINF_SB-2114: Vue 3 Fixed initial permalink navigation  and url parameter sync
With Vue 3 initial navigation to permalink did not work and the
application entered with the default settings instead with the one
provided by the query string of the permalink. This was only an issue
on initial application load, once the application was loaded we could
reenter the permalink to navigate to it.

This was due to event sequence that changed between Vue 2 and 3. To solve
this issue the application loading management and router store sync
management has been reworked to be less dependent on events sequence.

Note this also fixed a bug that was already present on Vue 2 where when
the application was loaded manually navigation to a specific layers
query param, leaving all other param empty, did not reroute to the
default other param; e.g. Navigate to /#/map?layers=my-layer did not reroute to
/#/map?layers=my-layer&lang=de&topic=ech&...