BGDIINF_SB-2120: Fixed scale-line not showing
If the initial zoom was below the threshold set in the template,
the scale-line would not show after zooming in past the threshold.

It seems as if a portal-target with multiple portals combined with
the lazy nature of v-if causes the target to get confused and mix
the markup/DOM of the different portals.

Using v-show instead of v-if fixes the issue.