BGDIINF_SB-2114: Fixed legacy param with external KML
When the legacy params contained an extern KML and a non default
language query, the KML was not displayed.

This was due to the language changes from the default to the non default
(e.g. from FR to EN). In this case we needs to reset the whole layers
config that contains the i18n translation, but the KML layer is not part
of the layer config.

This is now fixed for the kml layer by taking the previous active
config, but this is just a temporary workaround as it doesn't change the
layer i18n translation.