BGDIINF_SB-2114: Fixed e2e test for legacy params
The KML layer url can have two different urls, the first one is any url
passed in the url layer query param (could be any external kml service)
and the second one is the url generated by the drawing module (this url
point to our configured kml service). In e2e test if we use a fake url
in the layer param, then in the active layer we will have two kml
layers, the one from the fake service and the one from the drawing
module that point to the configured kml service.

To avoid this issue in e2e test we use the correct url in the e2e layers
query param. For this we need to replicate the Vue environment variable
in the Cypress env in order to know if the app is configured with the
dev or prod kml service.