BGDIINF_SB-2015: Added info-button to search results
We allow the user to add some layers through the search results. Analog to the layer selection/settings in the menu we want to provide a button that opens a modal with information about the layer. As is the case in the current production version.

This change adds that button and refactors/moves the modal used in the menu in order to avoid having a dependency between the search and the menu components.