BGDIINF_SB-2134: Re-enabled all tests and did minor refactoring
This commit removes the `.only` modifier from the new test.

In addition this changes `menuItems.get()` to `menuItems.toArray()`. Both methods lead to the same result but the latter makes the code easier to understand.

And finally this refactors the selector of the test layer into a variable (`testLayerSelector`) to make the code easier to read.