BGDIINF_SB-2133: Added e2e test for language switch
During the update to Vue3 a bug was fixed that reverted the configuration to its initial state and prevented the displayed language in the active layers and topic menu to change when the user switched the language in the settings.

The test changes the layer configuration after the initial load and checks if the configuration persists after changing the language.

This commit also introduces a new fake request so we can wait until the layers have been configured (@layers-configured).