BGDIINF_SB-2109: Fix missing anchor in drawings
There is a bug where the markers loose their anchor information when exported to KML. This results in the markers moving when closing and opening the drawing menu.

The problem seems to originate in the OpenLayers KML implementation (ol/format/KML). Specifically, the generation of the hotSpot tag.

This change introduces a workaround in the form of two helper functions that de-/serialize the anchor value separately.
THIS IS ONLY TEMPORARY until we manage to find a real fix and push it upstream.