BGDIINF_SB-2119: Replaced portal-vue with Vue teleport
As Vue3 comes with its own teleport feature we want to get rid of the portal-vue plugin. However, because the destination has to be mounted before the teleport and the order is on a first-come basis, this isn't a drop-in replacement.

This change adds a delay (through nextTick) before the teleports are attempted and to guarantee the correct order I added dedicated destinations for each teleport.