BGDIINF_SB-2170: Changed layer copyright position
The problem with the force-click in the test had already been addressed via a z-index. But that hides the copyright notice if the profile is present (minimized or not).

This change positions the copyright notice above the profile so that it is always in the bottom right-hand corner of the visible part of the map. That not only makes Cypress happy but also copyright holders.