BGDIINF_SB-2170: Moved store access to child component
The last commit added a new teleport which is not ideal.

Pascal had the idea to attach the copyright and profile components directly to the store instead of getting their data from the parent components. That way we will be able to move the components without having to use a teleport.

I like the idea and this commit changes the copyright component to get its data directly from the store. The commit also moves a CSS class (footer-visible) that I forgot in the last commit.