BGDIINF_SB-2179: Reorganized and updated footer components
We currently have the problem that we collect more and more teleports.

This commit reorganizes all footer related components (except the infobox and profile) and updates them to make use of Vuex instead of values provided by parent components.

With the reorganisation also came a new HTML structure and the CSS to go with it.