BGDIINF_SB-2179: Moved footer into map module and updated z-index
While managing the map instance through the Vuex store helps decouple the different components it could lead to performance problems because of the applied proxies.

Additionally, the map controls from the header (location and zoom) were on top of the profile view if the viewport wasn't high enough.

This commit moves the footer into the map module in order to use the injected getMap function and updates the z-index to put the profile on top.