BGDIINF_SB-2132: Changed text-add icon (drawing)
We currently use a plus icon for the text-add button which doesn't
communicate the purpose of the button as the other buttons from the
drawing toolbox also add things.

This changes the icon from a "+" to a "T".

More information about the choice can be found in the Jira ticket.