BGDIINF_SB-2227: Fixed w3w coords and added link
We had a mix-up with the coordinate values which resulted in
positions around Somalia. And one had to manually copy/paste the words
from the pop-up to the website to see the result.

This fixes (turns around) the coordinate values and
adds a link to the three words that leads to the website.