BGDIINF_SB-2226: Fixed position of share link
The link to share the current position points to the center of the
map. This is because the link uses the current URL as-is instead
of the ClickInfo position.

This replaces lat/lon when generating the URL so that the marker
is at the expected position when opening the link.