BGDIINF_SB-2237: Updated QR code link
During the review of BGDIINF_SB-2226 Brice noticed that the QR code
was missing the crosshair and that the link wouldn't update when
clicking on a different spot on the map.

With this update the QR code uses the same link as the "share link"
which also fixed the reactivity problem.

And while I was at it I also removed a redundant WGS84 format and
did some minor refactoring.