BGDIINF_SB-2245: Moved fixture to own file
As suggested by Pascal the fixture for the html-popup should be
in its own file. This makes the spec cleaner and allows for reuse.

I created a named export as there are different types of content
that we might want to use down the line.

This also includes some minor refactoring.