BGDIINF_SB-2228: Added copy-buttons and QR code
To keep the expected behavior for selecting text this removes the
user-select styles and adds copy-buttons for each format.

This also removes the toggle for the QR code and updates the size
of the overlay accordingly.

The short-links and the sizing of the title-bar will be done in
separate pull requests.