BGDIINF_SB-1388: Added preview for search results
In line with mf-geoadmin3 this adds a preview on hover to the search
results. The layers use the same functionality as the topic list and
the locations use the pinnedLocation that was already present.

This also includes some refactoring of the layer preview code and
adds the e2e tests for both the topic list and search results.