BGDIINF_SB-2323: Added condition to ignore faultily event
Currently, when someone tries to open the location pop-up on iOS
Safari the pop-up directly closes again. This is because the browser
sends a click event in addition to the context-menu event.

As there is nothing we can do against this event, I added a check
to the "singleclick" event to see if the last event was a context-
menu and if it was on the same position.