BGDIINF_SB-2361: Update code style
This mostly introduces some empty lines to give the
code some room to breathe and make it easier to scan.

I also moved the dragging setter/getter from the Feature
to the EditableFeature class and sorted the store modules
alphabetically.

And I fixed some English mistakes in the store ADR.