Fix after rebase
Also reactivating all profile.spec.js tests and fixing/removing part that are no longer relevant.
There's also a visibility test on profile-popup-tooltip that I had to remove because Cypress complains it is covered by the infobox header (with a position: fixed), which is not the case as the z-index makes it cover the header.