BGDIINF_SB-2235: Added ADR for mobile click
This is an ADR based on the JIRA ticket.

In order to enable users on mobile to also have access to the location popup, we have to think of a way to have 3 features on 2 gestures (touch and long touch).
The solution is to always run an "identify" request on short touch, and if nothing is found toggle the fullscreen mode. This frees the long click so that we can use it to show the location popup on mobile too.