Use of LV95 as projection for internal layers
In order to fix our "white border" issue, I've switched the default projection system we use to request tiles from our backend to LV95.
I've also declared a proper TileGrid for WMTS layers, which should cut the white part that was present on the side of tiles (OpenLayers will not require tiles outside of the extent, plus seems to crop images to fit this extent).
This way we should be ready to place the pixelkarte-farbe map on top of a vector tile layer, in order to achieve world wide coverage without compromising on details over Switzerland.