BGDIINF_SB-2611: add Swiss flag as favicon
Use a SVG favicon in the process instead of the old .ico file.
There's a fallback to a PNG image in case the browser doesn't support SVG favicons.
Those favicons were generated with the SVG in src/assets/svg/swiss-flag.svg using the online tool https://realfavicongenerator.net/

I've also set the title of the page to the official english title in the process, so that it may not accidentally show as "Vue app" anymore...