BGDIINF_SB-2611: applying latest standard for favicons
Found there : https://evilmartians.com/chronicles/how-to-favicon-in-2021-six-files-that-fit-most-needs
Sadly this means Chrome and Firefox will load more than one icon for their purposes. Usually they will load a PNG of their choosing and then the SVG, if the SVG is valid they will use it, otherwise the PNG version is used. There's no way around it if we want to serve our icon as SVG, there will always be a PNG load beforehand (at least for now)

This also adds a web manifest for our webapp. I've pointed to the swisstopo app in both apple and android store (even though in the end we will try to replace this app with our webapp :-) )