BGDIINF_SB-2608: Fixed missmatch between profile and geodesic line
When drawing a line or a measurement, we use a geodesic line. However the profile
backend did a flat filling to have at least 200 points. This lead to a missmatch
of a few pixels when drawing a line across switzerland.

To solve the issue we use the geodesic points as coordinate to pass to the
profile backend, this way the filling is done by the frontend instead of the
backend.