BGDIINF_SB-2472 : better handling of empty profile
As we still need to be able to delete a profile, even if empty, I've left the info container and the trash button. Otherwise, if the whole popup was hidden, the user was not able to delete an empty profile anymore.

In E2E test : adding a profile mockup function as it is used in 2 different files, and harmonisation of profile mockup usage.
Reactivation of tests that now works as intended.