Fixed button without icon text being not centered and fix open/close drawing menu styling
Some button without icons where using the ButtonWithIcon class and therefore
a margin to the label was added putting the text of center. To solve this
I made the icon required in the later class and changed all button that don't
have an icon to use directly bootstrap.

Also align the style of the drawing open/close menu to the one of the regular
menu.