BGDIINF_SB-2671: Put the assets in a versioned directory
This allow a better deployment approach on S3. We can copy first the versioned
assets directory (containing the images but also the javascript and css code)
into S3 without disturbing the current version, once it is copied, we
can copy the public files from the root and finally the index.html.

This procedure reduce the risk of downtime.