Fixed time selector button margin
This button uses the PopOverButton which uses the ButtonWithIcon which had
a fixed margin for the text. The issue is that the time selector button has
no icon, therefore we don't need any margin.
This has been temporarly fixed but we need to refactor the whole application
button (see BGDIINF_SB-2680)