BGDIINF_SB-2671: Fixed integration and production build
Production build did not used the production backends, while we did not
had integration build. This has been solved. This means that now
we can only deploy integration build from master on INT. When merging on master
we build the app for integration and for production.

The deploy script has been update to only deploy on dev, deploying on INT and
PROD will be done within another script.