BGDIINF_SB-2662: Location popup avoid new line on long what3words
Some what3words are long and cause the copy button to be pushed in a new line.
To avoid this and keep the copy button on the same line, we use the text overflow
ellipsis. To achieve this we also reduced the labels column width to a minimum
to get more space for the what3words. In case of long what3words the word
is now truncated with elipsis (...) but you can still copy the whole name using
the copy button. Most of the time the word is not truncated.