BGDIINF_SB-2683: Fixed GeoJSON layers
Some GeoJSON layers were not displayed correctly. Styling with a geometry type
of `point` where not displayed due to a wrong string comparison (`'point' === 'Point'`)

Also sometimes we could have a crash due to a non existing layer object which
was due to the fact that when hovering a search result a layer is created and
asynchronously the layer style was loaded, meanwhile when the hover is gone
the layer is destroyed which can happen before the layer style load from backend.

Updated some missing event in emit list to prevents vue warning.