BGDIINF_SB-2683: Fixed geojson layer feature selection
We could not select GeoJSON layer feature on the map. The issue was due to
wrong layerFilter in the OpenLayersMap.

Then once the feature selection was fixed the position was not correctly
marked as selected due to the wrong coordinates in the features.api were
the setter was not called.

Finally the content of the popup was wrong due to the missing feature name
in the LayerFeature.