BGDIINF_SB-2305 : replace SwissImage WMTS layer with VectorTile
There's an issue with that in the fact that our style.json doesn't define any boundary or extent with the embedded WMTS, and tiles are requested outside of Switzerland (resulting in many HTTP 404 responses).
Same with the PBF tiles, they are not more detailed outside of Switzerland, looking a bit weird without any satellite imagery on top.