BGDIINF_SB-2668: Fixed empty drawing being saved and displayed in menu
When closing the drawing menu without creating any drawing an empty kml was
created and added to the menu in the map displayed.

The saveKml event was called two many times, now it is only called when the
drawing effectively changes. I also renamed some function names to be more
explicit.