BGDIINF_SB-2019 : fix some bugs
Use of TippyJS with footer attribution disclaimer. Importing tippy style directly into the main App component and adding a style to have a red background/white text.

Fixing a bug now that external WMS do not use GeoAdminWMSLayer as base, the ID of the layer was not conveyed properly in OpenLayers components