BGDIINF_SB-2729: Fixed and improved save KML error handling
Before when the from the backend failed you couldn't close the drawing menu
and the error status was not displayed. Also when an error happened outside the
toggling of the show drawing menu, the exception was not catch.

Now we simply set the status to ERROR and re-trigger an update later on.