BGDIINF_SB-2729: Improved error handling and user feedback in drawing
Now if the loading of the KML failed in the drawing mode, we remove the
kml layer as it is not selectable and we print an error message. The loading
of the kml to translate it to a drawing will be automatically retried. In case
of error we also display an error message at the same place as for saving error.

Those errors message are also now displayed in red.

Also renamed the savingStatus to drawingState which refer more to a drawing state
to make the code clearer as it is also used to check if any modification on
the drawing has been done.

Improved the error handling and user experience in case of network failure.