Fixed Cypress drawing tests
Due to a too low number in the setTimout method we ended up sending several
requests in a row for saving a drawing instead of just 1. So increased
the timer to avoid such useless duplicates requests that slow down tests.

Also some request failed in Cypress due to a wrong interception fixture.