Fixed crash when running unit testing
When running the unit test at the end of the tests the generatebuildinfo plugin
generated a crash. Also this did not disturb the CI and the test were taken
as successful, the console output was garbaged with the crash output.

To avoid such crash the plugin directly exit if in test mode. It has also
been renamed to follow the vite plugin best practice.