BGDIINF_SB-2736: Fixed popover button (drawing styling and time selector) styling
The popover button uses tippy and its style has been broken due to the import
of the tippy css. Tippy css can only be imported globally to work. So in order
to not broke the popover button styling. I created a new tippy theme for it.

Also fixed the location popup tippy language change, before when the language
setting changed the tippy language was not updated.

Also use the light-border theme as default for tooltips.