BGDIINF_SB-2019: Fixed external layer and kml layer visibility flag in URL
When loading an External layer or a KML layer the visible flag in url in initial
load was not taken into account.

I found two issues although I did not fully understood the first second one.

- When adding a layer the visible flag was for an unknown reason hardcoded to true ?
I removed this hardcoding and make sure that the new and/or update layer config
is properly added to the active layers with its new current config.
- During the update store from query parameter, the visible flag was updated
using the toggleLayerVisibility() action. I did not fully understood why but in
some cases the toggle did not worked as expected and the final result was not
the one expected. This was probably due to some race conditions between events.
So to avoid such race conditions I explicitely used the setLayerVisibility()
action.

With this 2 fixes I could not find any other issue with the initial load of
external layer and kml layer regarding the visible flag and opacity.