BGDIINF_SB-2019: Fixed url param changes after initial load
After an initial load, changing manually the url (e.g. changing a layer visibilty
flag) did not triggered the action !

This was due to the last route.push() action being aborted due to duplicate
route and in this case the routeChangeIsTriggeredByThisModule flag was
not reseted and the next url change would have been ignored.