Fixed dropdown show not working after hide by click outside
When showing a dropdown button with toggle button and clicking outside of
the dropdown, bootstrap automatically hide the dropdown menu. However in this
case the `expanded` value of the component was not updated and a next click
on the component had no effect; the user had to click a second time to show
the dropdown menu.

To avoid this we let bootstrap to handle the `show` class. The `expanded` was
superfluous.