BGDIINF_SB-2305 : removing maplibre instance from reactive
Bad choice to add this instance as a computed, as it will then be reactive (meaning proxified and watched) and as MapLibre is not a simple object (many, many... attributes) it could get really slow (or really heavy on the CPU).
Also cleaning up imports for the OL vector layer component.