BGDIINF_SB-2728: Fixed legacy drawing initial load
On initial load of legacy drawing, it was not visible. To make it visible we
had to toggle the visible flag.

This was a race condition between the icons backend and the kml backend. In
order to display legacy drawings we need to have the icons sets information
from the icons backend. In our case the kml was loaded and added to openlayer
before the icons backend answered.

Now we update the drawing features on changes from the icons backend.