Adapt test to Cypress 12
Now Cypress forcefully clean the page before any test (to further improve test isolation), which is a good thing. Except for one of our test, the search bar coordinates one, which rely on a single cy.visit in order to improve performances (no need to reload the page for each coordinates entered in the search bar)

With this option, we find the previous behavior, which is that Cypress will keep the page opened as it was during the previous test.


There was also a last cy.server used here, which was removed in Cypress 12, I've replaced it with a cy.intercept