BGDIINF_SB-2733: Fixed layer cloning
Because the layers object are saved into the store, they are made reactive and
also cannot be changed outside a mutation. However there is some cases where
we want to change a copy outside a mutation in order to update the layer config
in the store. We do this with the clone() method. However this method did
not deeply cloned the objects.

In order to avoid using external library like lodash to do correct deep clone,
the clone method is overwritten in each subclass when needed.

This fix a future issue for the kml drawing admin id that needs to be removed
when exiting the drawing menu.