BGDIINF_SB-2733: Removed drawing adminId from url and open drawing mode automatically
Thee adminId is now removed from the URL and upon adminId detection, the drawing
menu is automatically opened.

Re-worked the KML metadata for simplification, now it is set together with
the layer in the layer store action. This simplify the url to layer store
logic.
Also avoid to have duplicate layer object in the store, now the active/selected
kml layer for drawing is only kept in the layer store, previously it was cloned
and also kept in the drawing store. The drawing menu has a copy of the layer
when the drawing mode is open, but this copy is only used by the drawing module
and is not in a store !

To avoid having the kml default opacity hardcoded in several places in the code
the KML layer class constructor has been changed in order to have the opacity
optional. So now the default opacity is only hardcoded in one place in the
KMLLayer class.