BGDIINF_SB-2744 : revert default ECH BG layer to given value
And add light base map behind pixelkarte-farbe, instead of switching to light base map by default.
Light base map is now added as an option in the BG selection wheel, and if chosen will be rendered in full (no exclusion over Switzerland).
The exclusion of data over Switzerland with LightBaseMap still occurs if it is used behind pixelkarte-farbe