BGDIINF_SB-2733: Keep drawing adminId in memory when leaving the drawing menu
The adminId is kept in memory until next reload. Also keep the layerId in the
url when inside the drawing mode in order not to loose if the page is reloaded.

We also uses the last KML in the active list (independently of visible flag) as
the kml to edit/copy. This is need because if the user is reloading the page
while editing the layer will be invisible after reloading, so the user still
have the possibility (without making the layer visible again) to copy edit
his drawing.