BGDIINF_SB-2733: Removing URL encoding hack
Let Vue router to handle URL encoding and parsing without a special hack
for the layers parameter. This caused troubled with the E2E tests.

Moreover we should be consistent in the encoding and not use any special case
for one parameter. This complicate the code and make it more error prone.