BGDIINF_SB-2751: Tried to make the language button more reactive
I tried to moved the language button in a component in order to decouple the
display of the language selection to make the selection feedback quicker, but this
unfortunately did not helped. The actions takes about 100ms due to the
vuex action.

Also did some small improvement in look and feel.