BGDIINF_SB-2751: Moved the language tool bar to top right corner on Desktop
Now the language tool bar is on top right corner on Desktop and we use link
button instead of regular buttons.

With this work the header styling has been reworked and cleaned up to use mainly
bootstrap classes. Also moved all CSS for styling inside the component and let
CSS positioning in the parent component. This make IMHO cleaner code as a component
is responsible for its look and feel and content disposition, while the parent
component is responsible for placement.

The search bar size has also been limited to 800px on desktop to avoid having
it filling the whole header.