BGDIINF_SB-2733: Fixed auto open drawing mode issue
When a KML admin ID was detected in the layer and the active layers was updated,
the drawing mode was automatically open. This was expected upon parsing the
KML admin ID from url, but this also happened when the active layers changed
like for example toggling any layer visibility or config or re-ordering layers.

Now we explicitely open the drawing mode on admin id parsed from url parsing.