BGDIINF_SB-2751: Fixed menu section scrolling
Because on desktop mode the settings section was removed from the menu, the
topics scrolling was brocken (the content was not anymore scrollable which
overflow on the acitve layer section).

This was because we used the `grid-template-rows` in css to set or unset the
scroll to each row and by removing a section we needed to change this css.
To avoid having a dependence between css and the number of section we now use
the `grid-auto-rows: auto` which makes every row scrollable. Because the main
section cannot be open simultaenously this is not an issue.