BGDIINF_SB-2733: Keep the kml layer visible flag to true while drawing
Now while drawing the kml layer visible flag is kept to true, this allow the
user to reload the page while drawing without loosing the visible flag. This
commit also fixed the issue that a drawing copy would be added to the active
layer list instead of replacing the copied drawing.

Re-order the layers store actions and mutation to have logic order, this
make the code easier to read/navigate.