BGDIINF_SB-2764 : add feedback form
that enables the user to rate his experience with a 5 star scale, and write his message directly in the UI (instead of letting him write it in his email editor).

Also fixing a bug on tablet layout where the links were not rendered properly in the menu/settings section, as the UI was still considered being in desktop mode, thus the feedback and more info buttons were still rendered as links.