BGDIINF_SB-2769 : add user feedback for wrong inputs
Give the user a visual feedback if the email he wrote is malformed. Only check format when the user has left the input, so that it doesn't start checking after first letter is typed in (and as a single letter is a malformed email it would directly give a bad feedback to the user who has not yet written anything...)

Also fixing a little bug where when closing the feedback modal by clicking on the transparent backdrop would not clear inputs.