BGDIINF_SB-2680 : remove ButtonWithIcon component
and replace all usage of this component with standard Bootstrap button.
Also simplifying PopoverButton component, so that the specialized version used by feature editing is stored there, and the simpler version is use as plain HTML/JS by the time selector.

Reworking the drawing toolbar so that it is more Bootrap based, and less custom CSS based.
Also replacing many instances of custom CSS code in component, that could be tackled by Bootstrap, with Bootstrap classes.