BGDIINF_SB-2784 : rework of the dev site warning placement
in order to enable the content of the header to be vertically centered by the flexbox.
Otherwise, as it was also present in the same container as the search bar, the height of the disclaimer was taken into account when placing the search bar input at the vertical center of the header.

This then covers what was intended for the ticket BGDIINF_SB-2259