BGDIINF_SB-2792 : fix layer ordering
There was a discrepancy between mf-geoadmin3 layer ordering (in the menu and URL) and web-mapviewer way of handling them.
In the URL, layers should be ordered from bottom to top, and on the side menu from top to bottom (layer "on top" of the layer stack is the top layer).
Topics were an issue as they list their layers to activate in the opposite order (top -> bottom) so I've inverted those layer before importing them.