BGDIINF_SB-2673 : switching from D3 to ChartJS for the profile chart
D3 starts to be a bit outdated and hard to handle with the segmentation approach.
ChartJS offers a wide variety of customization through plugin, and less direct canvas manipulation (as it functions how we want it to out of the box).
Adding two custom ChartJS module to handle the empty/nodata segments of the profile, and one to show the user the data model being used.
Using a zoom plugin so that users can now zoom in the profile to better see some feature of it.
Rework of the styling of the profile done in the process of switching to ChartJS (e.g. points are now barely visible, but can help to read/understand where the data is)