BGDIINF_SB-2673 : switch to in-memory CSV profile download
Instead of re-requesting data from the backend, we can create the CSV data on the fly as we've got everything we need.
Removing now unused API function to request profile as a CSV, and fixing the E2E test covering this part of the app.
Also fixing the tooltip positioning when hovering the profile chart.