BGDIINF_SB-2680 : switch back to store isDesktopMode/isPhoneMode
for drawing toolbox, as it was doing some weird breakpoint interpretation when holding the phone on landscape mode, switching to the desktop layout of the toolbox without completely doing so.

So I removed all the logic that was switching based on Bootstrap breakpoint and replaced it with v-if or v-class based approach based on the store values.