BGDIINF_SB-2743 : aggregate TippyJS usage for third party disclaimer
There was sometimes issue with TippyJS tooltip not showing up in the footer attributions. As it was kind of the same usage of TippyJS as the one found in the active layer menu, I aggregated the use into one single component, which encapsulate all the TippyJS logic.
This should fix the discrepancy between the two use cases.