BGDIINF_SB-2829: Text in tooltips is now selectable
Text in the tooltips of features was previously not selectable. This
commit fixes this. The classname map-popover-overlay did not exist.
Removing this classname activates instead the default class name
ol-selectable and ol-overlay-container which makes the text in the
overlay selectable. (The overlay is a child of the openlayersmap which
deactivates selection)